[KSM] Git ignore update & App.js
Git ignore update & App.js
Add file directory
