+++ myhealing_client/src/index.js
@@ -3,15 +3,10 @@ import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
-import Inform from './Inform';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
-  <React.StrictMode>
-    <Inform text="지도로 맞춤 장소 찾기"/>
-    <App />
-    <Inform text="키워드로 맞춤 장소 찾기"/>
-  </React.StrictMode>
+  <App />
 );
 
 // If you want to start measuring performance in your app, pass a function
