+++ myhealing_client/src/page/bottom/Bottom.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+function Bottom() {
+  return (
+    <h3>Bottom 입니다.</h3>
+  );
+}
+
+export default Bottom;
