+++ myhealing_client/src/page/header/Header.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+function Header() {
+  return (
+    <h3>header 입니다.</h3>
+  );
+}
+
+export default Header;
