+++ myhealing_client/src/page/map/Inform.js
